Move shared helpers to separate file
