+++ client/go/cmd/helpers.go
@@ -1,6 +1,6 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-// Models a target for Vespa commands
-// author: bratseth
+// Helpers used by multiple sub-commands.
+// Author: mpolden
 
 package cmd
 
@@ -8,12 +8,66 @@ import (
 	"crypto/tls"
 	"fmt"
 	"log"
+	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
 	"github.com/vespa-engine/vespa/vespa"
 )
+
+var exitFunc = os.Exit // To allow overriding Exit in tests
+
+func printErrHint(err error, hints ...string) {
+	printErr(nil, err.Error())
+	for _, hint := range hints {
+		log.Print(color.Cyan("Hint: "), hint)
+	}
+	exitFunc(1)
+}
+
+func printErr(err error, msg ...interface{}) {
+	if len(msg) > 0 {
+		log.Print(color.Red("Error: "), fmt.Sprint(msg...))
+	}
+	if err != nil {
+		log.Print(color.Yellow(err))
+	}
+	exitFunc(1)
+}
+
+func printSuccess(msg ...interface{}) {
+	log.Print(color.Green("Success: "), fmt.Sprint(msg...))
+}
+
+func readAPIKey(tenant string) []byte {
+	configDir := configDir("")
+	apiKeyPath := filepath.Join(configDir, tenant+".api-key.pem")
+	key, err := os.ReadFile(apiKeyPath)
+	if err != nil {
+		printErrHint(err, "Deployment to cloud requires an API key. Try 'vespa api-key'")
+	}
+	return key
+}
+
+func deploymentFromArgs() vespa.Deployment {
+	zone, err := vespa.ZoneFromString(zoneArg)
+	if err != nil {
+		printErrHint(err, "Zone format is <env>.<region>")
+	}
+	app, err := vespa.ApplicationFromString(getApplication())
+	if err != nil {
+		printErrHint(err, "Application format is <tenant>.<app>.<instance>")
+	}
+	return vespa.Deployment{Application: app, Zone: zone}
+}
+
+func applicationSource(args []string) string {
+	if len(args) > 0 {
+		return args[0]
+	}
+	return "."
+}
 
 func getApplication() string {
 	app, err := getOption(applicationFlag)
